added stuff
added file picker for the sheet
cancel
title and author name sync between types
sheet showing when there is one result
trending not refreshing every time app shows

+++ app/src/main/java/com/deniscerri/ytdlnis/database/dao/DownloadDao.kt
@@ -22,7 +22,7 @@ interface DownloadDao {
     fun getDownloadByWorkId(workID: Int) : DownloadItem
 
     @Query("SELECT * FROM downloads WHERE id=:id LIMIT 1")
-    fun getDownloadById(id: Int) : DownloadItem
+    fun getDownloadById(id: Long) : DownloadItem
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(item: DownloadItem) : Long
